Issue #792: Move algebraic datatype MappingOutcome to own package to enforce visibility.
Concrete implementations of MappingOutcome should not be visible elsewhere.
