+++ services/connectivity/messaging/src/main/java/org/eclipse/ditto/services/connectivity/messaging/mappingoutcome/DroppedOutcome.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright (c) 2020 Contributors to the Eclipse Foundation
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ */
+package org.eclipse.ditto.services.connectivity.messaging.mappingoutcome;
+
+import javax.annotation.Nullable;
+
+import org.eclipse.ditto.services.models.connectivity.ExternalMessage;
+
+// private to MappingOutcome. Do NOT use directly.
+final class DroppedOutcome<T> implements MappingOutcome<T> {
+
+    private final ExternalMessage droppedMessage;
+
+    DroppedOutcome(@Nullable final ExternalMessage droppedMessage) {
+        this.droppedMessage = droppedMessage;
+    }
+
+    @Override
+    public <R> R accept(final Visitor<T, R> visitor) {
+        try {
+            return visitor.onDropped(droppedMessage);
+        } catch (final Exception e) {
+            return visitor.onError(e, null, droppedMessage);
+        }
+    }
+}
