+++ services/connectivity/messaging/src/main/java/org/eclipse/ditto/services/connectivity/messaging/mappingoutcome/package-info.java
@@ -1,0 +1,20 @@
+/*
+ * Copyright (c) 2020 Contributors to the Eclipse Foundation
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ */
+
+
+/**
+ * This package is reserved for the algebraic data type {@code MappingOutcome}
+ * so that concrete implementations of {@code MappingOutcome} are not visible elsewhere.
+ */
+@org.eclipse.ditto.utils.jsr305.annotations.AllParametersAndReturnValuesAreNonnullByDefault
+package org.eclipse.ditto.services.connectivity.messaging.mappingoutcome;
